+++ src/elements/numberformat/index.ts
@@ -1,12 +1,16 @@
 import HTMLIntlNumberFormatElement from './numberformat';
 import HTMLIntlNumberFormatFormatElement from './numberformat-format';
 import HTMLIntlNumberFormatFormatToPartsElement from './numberformat-formattoparts';
+// import HTMLIntlNumberFormatFormatRangeElement from './numberformat-formatrange';
+// import HTMLIntlNumberFormatFormatRangeToPartsElement from './numberformat-formatrangetoparts';
 
 declare global {
   interface HTMLElementTagNameMap {
     'intl-numberformat': HTMLIntlNumberFormatElement;
     'intl-numberformat-format': HTMLIntlNumberFormatFormatElement;
     'intl-numberformat-formattoparts': HTMLIntlNumberFormatFormatToPartsElement;
+    // 'intl-numberformat-formatrange': HTMLIntlNumberFormatFormatRangeElement;
+    // 'intl-numberformat-formatrangetoparts': HTMLIntlNumberFormatFormatRangeToPartsElement;
   }
 }
 
@@ -14,10 +18,14 @@ export function defineIntlNumberFormatElements() {
   customElements.define('intl-numberformat', HTMLIntlNumberFormatElement);
   customElements.define('intl-numberformat-format', HTMLIntlNumberFormatFormatElement);
   customElements.define('intl-numberformat-formattoparts', HTMLIntlNumberFormatFormatToPartsElement);
+  // customElements.define('intl-numberformat-formatrange', HTMLIntlNumberFormatFormatRangeElement);
+  // customElements.define('intl-numberformat-formatrangetoparts', HTMLIntlNumberFormatFormatRangeToPartsElement);
 }
 
 export {
   HTMLIntlNumberFormatElement,
   HTMLIntlNumberFormatFormatElement,
   HTMLIntlNumberFormatFormatToPartsElement,
+  // HTMLIntlNumberFormatFormatRangeElement,
+  // HTMLIntlNumberFormatFormatRangeToPartsElement,
 }